+++ cmd/gpud/command/run.go
@@ -9,18 +9,15 @@ import (
 	"runtime"
 	"time"
 
-	sd "github.com/coreos/go-systemd/v22/daemon"
 	"github.com/gin-gonic/gin"
 	"github.com/urfave/cli"
 	"go.uber.org/zap"
-	"golang.org/x/sys/unix"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	"github.com/leptonai/gpud/pkg/config"
 	gpud_manager "github.com/leptonai/gpud/pkg/gpud-manager"
 	"github.com/leptonai/gpud/pkg/log"
 	gpudserver "github.com/leptonai/gpud/pkg/server"
-	"github.com/leptonai/gpud/pkg/systemd"
 	pkgsystemd "github.com/leptonai/gpud/pkg/systemd"
 	"github.com/leptonai/gpud/version"
 )
@@ -96,8 +93,8 @@ func cmdRun(cliContext *cli.Context) error {
 	log.Logger.Infof("starting gpud %v", version.Version)
 
 	done := gpudserver.HandleSignals(rootCtx, rootCancel, signals, serverC, func(ctx context.Context) error {
-		if systemd.SystemctlExists() {
-			if err := notifyStopping(ctx); err != nil {
+		if pkgsystemd.SystemctlExists() {
+			if err := pkgsystemd.NotifyStopping(ctx); err != nil {
 				log.Logger.Errorw("notify stopping failed")
 			}
 		}
@@ -106,7 +103,7 @@ func cmdRun(cliContext *cli.Context) error {
 
 	// start the signal handler as soon as we can to make sure that
 	// we don't miss any signals during boot
-	signal.Notify(signals, handledSignals...)
+	signal.Notify(signals, gpudserver.DefaultSignalsToHandle...)
 	m, err := gpud_manager.New()
 	if err != nil {
 		return err
@@ -120,7 +117,7 @@ func cmdRun(cliContext *cli.Context) error {
 	serverC <- server
 
 	if pkgsystemd.SystemctlExists() {
-		if err := notifyReady(rootCtx); err != nil {
+		if err := pkgsystemd.NotifyReady(rootCtx); err != nil {
 			log.Logger.Warnw("notify ready failed")
 		}
 	} else {
@@ -131,27 +128,4 @@ func cmdRun(cliContext *cli.Context) error {
 	<-done
 
 	return nil
-}
-
-var handledSignals = []os.Signal{
-	unix.SIGTERM,
-	unix.SIGINT,
-	unix.SIGUSR1,
-	unix.SIGPIPE,
-}
-
-// notifyReady notifies systemd that the daemon is ready to serve requests
-func notifyReady(_ context.Context) error {
-	return sdNotify(sd.SdNotifyReady)
-}
-
-// notifyStopping notifies systemd that the daemon is about to be stopped
-func notifyStopping(_ context.Context) error {
-	return sdNotify(sd.SdNotifyStopping)
-}
-
-func sdNotify(state string) error {
-	notified, err := sd.SdNotify(false, state)
-	log.Logger.Debugf("sd notification: %v %v %v", state, notified, err)
-	return err
 }
